added confirmation

+++ cmd/kronk/remove/remove.go
@@ -24,12 +24,21 @@ func Run(args []string) error {
 		return err
 	}
 
+	modelFileName := filepath.Base(modelFile)
+
+	fmt.Printf("\nAre you sure you want to remove %q? (y/n): ", modelFileName)
+	var response string
+	fmt.Scanln(&response)
+	if response != "y" && response != "Y" {
+		fmt.Println("Remove cancelled")
+		return nil
+	}
+
 	if err := os.Remove(modelFile); err != nil {
 		return fmt.Errorf("unable to remove %q", modelFile)
 	}
 
 	// This file may not exist, so deleting it blindly.
-	modelFileName := filepath.Base(modelFile)
 	projFileName := fmt.Sprintf("mmproj-%s", modelFileName)
 	projFile := strings.Replace(modelFile, modelFileName, projFileName, 1)
 	os.Remove(projFile)
